First rough decouple of MDW behaviour
